style(shlvl_init): norm対応、ファイル切り分け
